# apps/api/internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: М3105 API
  version: "1.0"
  summary: Публичное API сайта группы М3105
  description: |
    Публичные эндпоинты сайта группы: предметы, лабораторные с дедлайнами, календарь и ICS-фид,
    конспекты, квизы, ЧаВо, статические страницы, поиск и расписание сдач.

    Все ответы в JSON (кроме `/calendar.ics` и файлов `/files/*`), время в RFC 3339 с таймзоной, даты без времени в формате `YYYY-MM-DD`.
    Возвращаются только опубликованные материалы. Авторизация для этих эндпоинтов не нужна.

    Эндпоинты админки (`/admin/*`) и аккаунтов студентов (`/auth/*`, `/me/*`, запись на сдачи)
    работают по HttpOnly-кукам и описаны в `docs/api.md` репозитория.
  contact:
    url: https://m3105.ru
security: []  # публичные эндпоинты без авторизации
servers:
  - url: https://m3105.ru/api/v1
    description: Продакшен
  - url: http://localhost:8080/api/v1
    description: Локальный API
...
paths:
...
  /settings:
    get:
      tags: [site]
      summary: Настройки сайта, текущая неделя и навигация
      operationId: getSettings
      responses:
        "200":
          description: Настройки
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SettingsResponse" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /home:
    get:
      tags: [site]
      summary: Сводка для главной страницы
      description: Ближайшие и просроченные дедлайны лаб, события и сдачи, свежие конспекты, предметы.
      operationId: getHome
      responses:
        "200":
          description: Сводка
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HomeResponse" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /subjects:
    get:
      tags: [subjects]
      summary: Список предметов
      description: Предметы в порядке `position` со счётчиками опубликованных лаб, конспектов и квизов.
      operationId: listSubjects
      responses:
        "200":
          description: Предметы
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SubjectWithCounts" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /subjects/{slug}:
    get:
      tags: [subjects]
      summary: Предмет с его лабами, конспектами и квизами
      operationId: getSubject
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Предмет
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubjectResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /labs:
    get:
      tags: [labs]
      summary: Список лабораторных
      description: Опубликованные лабы всех предметов или одного предмета. Длинные MDX-поля (`content`, `requirements`, `submission`, `variants`) в списке пустые.
      operationId: listLabs
      parameters:
        - $ref: "#/components/parameters/SubjectFilter"
      responses:
        "200":
          description: Лабы
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Lab" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /labs/{subject}/{slug}:
    get:
      tags: [labs]
      summary: Лабораторная целиком
      operationId: getLab
      parameters:
        - $ref: "#/components/parameters/SubjectSlug"
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Лаба
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lab" }
        "404": { $ref: "#/components/responses/NotFound" }
  /calendar:
    get:
      tags: [calendar]
      summary: Календарь за период
      description: |
        Дедлайны лабораторных, события (контрольные, экзамены, консультации) и сдачи.
        Поле `source` различает `lab`, `event` и `practice`. Без параметров возвращает окно от месяца назад до полугода вперёд.
        `from` и `to` принимают RFC 3339 или дату `YYYY-MM-DD` (дата в `to` включает весь день).
        Некорректное значение молча заменяется значением по умолчанию.
      operationId: getCalendar
      parameters:
        - name: from
          in: query
          schema: { type: string, examples: ["2026-09-01", "2026-09-01T00:00:00+03:00"] }
          description: Начало периода
        - name: to
          in: query
          schema: { type: string, examples: ["2026-12-31"] }
          description: Конец периода
        - $ref: "#/components/parameters/SubjectFilter"
      responses:
        "200":
          description: Элементы календаря
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CalendarResponse" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /calendar.ics:
    get:
      tags: [calendar]
      summary: iCalendar-фид календаря
      description: |
        Подписка для Google Calendar, Apple Calendar и Outlook: дедлайны лаб, события и сдачи. Окно: два месяца назад и год вперёд.
        У дедлайнов, контрольных и экзаменов напоминание за сутки. Ответ кэшируется на 15 минут.
      operationId: getCalendarICS
      parameters:
        - $ref: "#/components/parameters/SubjectFilter"
      responses:
        "200":
          description: Файл календаря
          headers:
            Content-Disposition:
              schema: { type: string, examples: ['inline; filename="m3105.ics"'] }
            Cache-Control:
              schema: { type: string, examples: ["public, max-age=900"] }
          content:
            text/calendar:
              schema: { type: string }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /notes:
    get:
      tags: [notes]
      summary: Список конспектов
      description: Опубликованные конспекты, по всем предметам или по одному. Поле `content` в списке пустое.
      operationId: listNotes
      parameters:
        - $ref: "#/components/parameters/SubjectFilter"
      responses:
        "200":
          description: Конспекты
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Note" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /notes/{subject}/{slug}:
    get:
      tags: [notes]
      summary: Конспект целиком
      description: Вместе с привязанными квизами и соседними конспектами предмета для навигации.
      operationId: getNote
      parameters:
        - $ref: "#/components/parameters/SubjectSlug"
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Конспект
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /quizzes:
    get:
      tags: [quizzes]
      summary: Список квизов
      operationId: listQuizzes
      parameters:
        - $ref: "#/components/parameters/SubjectFilter"
      responses:
        "200":
          description: Квизы
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/QuizSummary" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /quizzes/{slug}:
    get:
      tags: [quizzes]
      summary: Квиз с вопросами
      description: Формат вопросов описан в `docs/quiz-format.md`. Правильные ответы включены, проверка выполняется на клиенте.
      operationId: getQuiz
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Квиз
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quiz" }
        "404": { $ref: "#/components/responses/NotFound" }
  /faq:
    get:
      tags: [content]
      summary: Частые вопросы
      operationId: listFAQ
      responses:
        "200":
          description: Вопросы в порядке `position`
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FAQItem" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /pages:
    get:
      tags: [content]
      summary: Статические страницы
      operationId: listPages
      responses:
        "200":
          description: Опубликованные страницы, `content` в списке пустой
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Page" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /pages/{slug}:
    get:
      tags: [content]
      summary: Статическая страница
      operationId: getPage
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Страница
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Page" }
        "404": { $ref: "#/components/responses/NotFound" }
  /search:
    get:
      tags: [search]
      summary: Поиск по опубликованному
      description: |
        Полнотекстовый поиск по лабам, конспектам, квизам, ЧаВо, страницам и предметам.
        Русский словарь с учётом словоформ, последнее слово ищется по префиксу (поиск по мере набора),
        опечатки ловятся триграммами. Результаты отсортированы по релевантности: совпадение в заголовке
        весит больше, чем в тексте. Пустой запрос и запрос без букв и цифр дают пустой список.
      operationId: search
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, examples: ["указатели"] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
          description: Значения вне диапазона заменяются на 20
      responses:
        "200":
          description: Результаты
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SearchResult" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /practice:
    get:
      tags: [practice]
      summary: Занятия для сдачи лаб
      description: |
        По умолчанию только будущие. Для анонимных запросов `participants`, `myLabIds` и `availableLabs` пустые,
        `signedIn` равно `false`; вошедшие студенты видят участников и могут записаться (см. `docs/api.md`).
      operationId: listPractice
      parameters:
        - $ref: "#/components/parameters/SubjectFilter"
        - name: past
          in: query
          schema: { type: string, examples: ["1"] }
          description: Любое непустое значение включает прошедшие занятия
      responses:
        "200":
          description: Занятия
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PracticeListResponse" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /practice/{id}:
    get:
      tags: [practice]
      summary: Одно занятие
      operationId: getPractice
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Занятие
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PracticeSessionView" }
        "404": { $ref: "#/components/responses/NotFound" }
  /comments/{target}/{id}:
    get:
      tags: [social]
      summary: Комментарии к конспекту, лабе или посту
      description: |
        Читать может кто угодно. `POST` с тем же путём и телом `{ body, attachmentIds? }` требует студенческую куку `edu_user`,
        `DELETE /comments/{id}` удаляет свой комментарий. Конспект или лаба должны быть опубликованы.
        `attachmentIds` — id своих файлов из `POST /files`, ещё ни к чему не прикреплённых; с файлами текст можно не писать.
      operationId: listComments
      parameters:
        - name: target
          in: path
          required: true
          schema: { type: string, enum: [note, lab, post] }
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Комментарии, старые первыми
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Comment" }
        "404": { $ref: "#/components/responses/NotFound" }
  /posts:
    get:
      tags: [social]
      summary: Лента постов
      description: |
        Пользовательские посты одной ленты. `POST /posts` с `PostInput` создаёт пост от имени студента,
        `PUT`/`DELETE /posts/{id}` меняют и удаляют свой пост, `PUT`/`DELETE /posts/{id}/like` ставят и снимают лайк.
        Поля `liked` и `mine` считаются для студента, если передана кука `edu_user`.
        Без куки отдаются только посты с `visibility: public`: `members` видят лишь вошедшие студенты,
        и посты с `nsfw: true` всегда `members`. Комментарии к скрытому посту анонимам тоже недоступны (404),
        как и файлы такого поста.
      operationId: listPosts
      parameters:
        - name: kind
          in: query
          required: true
          schema: { type: string, enum: [shawarma, joke] }
        - name: sort
          in: query
          schema: { type: string, enum: [new, top], default: new }
          description: "`top` — сначала посты с наибольшим числом лайков"
      responses:
        "200":
          description: Посты
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Post" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /files/{id}/{name}:
    get:
      tags: [files]
      summary: Файл вложения
      description: |
        Адрес файла приходит в поле `url` у `Attachment`; последний сегмент — транслит имени, сервер смотрит только на `id`.
        Картинки, PDF, видео и аудио открываются в браузере, остальное скачивается; `?download=1` скачивает всегда.
        `?w=160|320|640|1280|1920` у JPEG, PNG и WebP отдаёт уменьшенную копию (WebP, если он есть в `Accept`, `Vary: Accept`);
        картинка уже нужной ширины, GIF, SVG и остальные типы приходят оригиналом.
        Поддерживаются `Range` и `ETag`. Файлы поста «для своих» анонимам не отдаются (404),
        как и загрузки, ещё не прикреплённые к комментарию или посту, — их видит только автор.

        Загрузка: `POST /files` (студент, кука `edu_user`, картинки JPG/PNG/GIF/WebP и PDF до 10 МБ, 30 файлов за 12 часов)
        и `POST /admin/files` (админка, любые файлы до 50 МБ), тело `multipart/form-data` с полем `file`, ответ — `Attachment`.
        У фото вырезаются EXIF и прочие метаданные, ориентация сохраняется.
      operationId: getFile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, pattern: "^[a-z2-7]{26}$" }
        - name: name
          in: path
          required: true
          schema: { type: string }
        - name: download
          in: query
          schema: { type: string, enum: ["1"] }
          description: Отдать как вложение, а не открыть в браузере
        - name: w
          in: query
          schema: { type: integer, enum: [160, 320, 640, 1280, 1920] }
          description: Ширина превью картинки; другие значения игнорируются
      responses:
        "200":
          description: Содержимое файла
          headers:
            Content-Disposition:
              schema: { type: string, examples: ["inline; filename=photo.jpg"] }
            Cache-Control:
              schema: { type: string, examples: ["public, max-age=31536000, immutable"] }
          content:
            "*/*":
              schema: { type: string, format: binary }
        "206":
          description: Часть файла по заголовку `Range`
        "404": { $ref: "#/components/responses/NotFound" }
components:
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string, pattern: "^[a-z0-9-]+$" }
      description: Слаг, сгенерированный из названия транслитерацией
    SubjectSlug:
      name: subject
      in: path
      required: true
      schema: { type: string, pattern: "^[a-z0-9-]+$" }
      description: Слаг предмета
    SubjectFilter:
      name: subject
      in: query
      schema: { type: string }
      description: Слаг предмета для фильтрации
  responses:
    ClientError:
      description: Некорректный запрос или превышен лимит; формат ошибки одинаковый для всех кодов
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Не найдено или не опубликовано
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Не найдено", code: not_found }
  schemas:
...
    SubjectWithCounts:
      allOf:
        - $ref: "#/components/schemas/Subject"
        - type: object
          properties:
            labsCount: { type: integer }
            notesCount: { type: integer }
            quizzesCount: { type: integer }
          required: [labsCount, notesCount, quizzesCount]
...
    Lab:
      allOf:
        - $ref: "#/components/schemas/SubjectJoin"
        - type: object
          properties:
            id: { type: integer }
            subjectId: { type: integer }
            number: { type: integer, minimum: 0, description: "Ноль допустим для необязательной вводной лабы" }
            slug: { type: string }
            title: { type: string }
            summary: { type: string }
            content: { type: string, description: "Задание, MDX" }
            requirements: { type: string, description: "Требования, MDX" }
            submission: { type: string, description: "Как сдавать, MDX" }
            variants: { type: string, description: "Варианты, MDX" }
            materials:
              type: array
              items: { $ref: "#/components/schemas/Link" }
            deadlineAt: { oneOf: [{ $ref: "#/components/schemas/DateTime" }, { type: "null" }] }
            deadlineNote: { type: string }
            maxScore: { oneOf: [{ type: integer }, { type: "null" }] }
            teacher: { type: string }
            status: { $ref: "#/components/schemas/Status" }
            createdAt: { $ref: "#/components/schemas/DateTime" }
            updatedAt: { $ref: "#/components/schemas/DateTime" }
          required: [id, subjectId, number, slug, title, summary, content, requirements, submission, variants, materials, deadlineAt, deadlineNote, maxScore, teacher, status, createdAt, updatedAt]
    Note:
      allOf:
        - $ref: "#/components/schemas/SubjectJoin"
        - type: object
          properties:
            id: { type: integer }
            subjectId: { type: integer }
            number: { type: integer, minimum: 1 }
            slug: { type: string }
            title: { type: string }
            summary: { type: string }
            content: { type: string, description: "MDX с формулами и кодом" }
            lectureDate: { oneOf: [{ $ref: "#/components/schemas/Date" }, { type: "null" }] }
            status: { $ref: "#/components/schemas/Status" }
            createdAt: { $ref: "#/components/schemas/DateTime" }
            updatedAt: { $ref: "#/components/schemas/DateTime" }
            quizzesCount: { type: integer, description: "Опубликованные квизы, привязанные к конспекту" }
          required: [id, subjectId, number, slug, title, summary, content, lectureDate, status, createdAt, updatedAt, quizzesCount]
    NoteResponse:
      allOf:
        - $ref: "#/components/schemas/Note"
        - type: object
          properties:
            quizzes:
              type: array
              items: { $ref: "#/components/schemas/QuizSummary" }
            prev: { oneOf: [{ $ref: "#/components/schemas/Note" }, { type: "null" }] }
            next: { oneOf: [{ $ref: "#/components/schemas/Note" }, { type: "null" }] }
          required: [quizzes, prev, next]
...
    Quiz:
      allOf:
        - $ref: "#/components/schemas/SubjectJoin"
        - type: object
          properties:
            id: { type: integer }
            subjectId: { oneOf: [{ type: integer }, { type: "null" }] }
            noteId: { oneOf: [{ type: integer }, { type: "null" }] }
            slug: { type: string }
            title: { type: string }
            description: { type: string }
            questions:
              type: array
              items: { $ref: "#/components/schemas/QuizQuestion" }
            shuffleQuestions: { type: boolean }
            shuffleOptions: { type: boolean }
            status: { $ref: "#/components/schemas/Status" }
            createdAt: { $ref: "#/components/schemas/DateTime" }
            updatedAt: { $ref: "#/components/schemas/DateTime" }
            noteSlug: { type: string }
            noteTitle: { type: string }
          required: [id, subjectId, noteId, slug, title, description, questions, shuffleQuestions, shuffleOptions, status, createdAt, updatedAt, noteSlug, noteTitle]
    QuizSummary:
      allOf:
        - $ref: "#/components/schemas/Quiz"
        - type: object
          properties:
            questionsCount: { type: integer }
          required: [questionsCount]
    FAQItem:
      type: object
      properties:
        id: { type: integer }
        question: { type: string }
        answer: { type: string, description: "MDX" }
        category: { type: string }
        position: { type: integer }
        createdAt: { $ref: "#/components/schemas/DateTime" }
        updatedAt: { $ref: "#/components/schemas/DateTime" }
      required: [id, question, answer, category, position, createdAt, updatedAt]
...
    SearchResult:
      type: object
      properties:
        kind:
          type: string
          enum: [lab, note, quiz, faq, page, subject]
        title: { type: string }
        subtitle: { type: string }
        path: { type: string, description: "Путь на сайте" }
        snippet: { type: string, description: "Фрагмент текста, совпадения обёрнуты в <mark>; может быть пустым" }
      required: [kind, title, subtitle, path, snippet]
    SettingsResponse:
      type: object
      properties:
        settings: { $ref: "#/components/schemas/Settings" }
        week: { $ref: "#/components/schemas/Week" }
        now: { $ref: "#/components/schemas/DateTime" }
        navPages:
          type: array
          items: { $ref: "#/components/schemas/Page" }
          description: Страницы с `showInNav`, без `content`
        auth: { $ref: "#/components/schemas/AuthInfo" }
      required: [settings, week, now, navPages, auth]
    HomeResponse:
      type: object
      properties:
        settings: { $ref: "#/components/schemas/Settings" }
        week: { $ref: "#/components/schemas/Week" }
        now: { $ref: "#/components/schemas/DateTime" }
        upcomingDeadlines:
          type: array
          items: { $ref: "#/components/schemas/CalendarItem" }
          description: Только дедлайны лаб
        overdueDeadlines:
          type: array
          items: { $ref: "#/components/schemas/CalendarItem" }
        upcomingEvents:
          type: array
          items: { $ref: "#/components/schemas/CalendarItem" }
          description: Контрольные, экзамены, консультации и сдачи
        recentNotes:
          type: array
          items: { $ref: "#/components/schemas/Note" }
        subjects:
          type: array
          items: { $ref: "#/components/schemas/SubjectWithCounts" }
      required: [settings, week, now, upcomingDeadlines, overdueDeadlines, upcomingEvents, recentNotes, subjects]
    SubjectResponse:
      type: object
      properties:
        subject: { $ref: "#/components/schemas/Subject" }
        labs:
          type: array
          items: { $ref: "#/components/schemas/Lab" }
        notes:
          type: array
          items: { $ref: "#/components/schemas/Note" }
        quizzes:
          type: array
          items: { $ref: "#/components/schemas/QuizSummary" }
      required: [subject, labs, notes, quizzes]
    CalendarResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CalendarItem" }
        from: { $ref: "#/components/schemas/DateTime" }
        to: { $ref: "#/components/schemas/DateTime" }
        now: { $ref: "#/components/schemas/DateTime" }
      required: [items, from, to, now]
...